app/game_data_service/requirements.txt
anyio==3.6.2
cassandra-driver==3.26.0
click==8.1.3
fastapi==0.95.1
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
pydantic==1.10.7
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
